\name{getLevels}
\alias{getLevels}
\alias{getLevels,CuffSet-method}
\alias{getLevels,CuffData-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getLevels
}
\description{
Returns a list of samples as levels. This should not be called directly by user.
}
\usage{
\S4method{getLevels}{CuffData}(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class 'CuffData' or 'CuffFeatureSet' or 'CuffFeature'
}
}
\details{
For internal usage only.
}
\value{
A vector of sample names as factors.
}
\references{
None.
}
\author{
Loyal A. Goff
}
\note{
None.
}
