\name{getRepLevels}
\alias{getRepLevels}
\alias{getRepLevels,CuffSet-method}
\alias{getRepLevels,CuffData-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getRepLevels
}
\description{
Returns a list of replicate samples as levels. This should not be called directly by user.
}
\usage{
\S4method{getRepLevels}{CuffData}(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class 'CuffSet' or 'CuffData'
}
}
\details{
For internal usage only.
}
\value{
A vector of replicate names as factors.
}
\references{
None.
}
\author{
Loyal A. Goff
}
\note{
None.
}
