\name{PINK1}
\alias{PINK1}
\docType{data}
\title{
PINK1
}
\description{
A sample 'CuffGene' dataset
}
\usage{data(sampleData)}
\format{
  PINK1 is a CuffGene object (extends CuffFeature) with all sample gene-, isoform-, TSS-, and CDS-level data for the gene 'PINK1'. 
}
\details{
Sample CuffGene data for gene 'PINK1'
}
\source{
None
}
\references{
None
}
\examples{
data(sampleData)
PINK1
}
\keyword{datasets}
