\name{sampleGeneSet}
\alias{sampleGeneSet}
\docType{data}
\title{
sampleGeneSet
}
\description{
A sample CuffGeneSet data set for 20 genes.
}
\usage{data(sampleData)}
\format{
  sampleGeneSet is a CuffGeneSet (extends CuffFeatureSet) object containing all sample gene-, isoform-, TSS-, and CDS-level data for 20 different genes.
  These data were derived from a toy set of hESC-vs-iPSC-vs-Fibroblast RNA-Seq expression data.
}
\details{
None
}
\source{
None
}
\references{
None
}
\examples{
data(sampleData)

}
\keyword{datasets}
