\name{sampleIDs}
\alias{sampleIDs}
\docType{data}
\title{
sampleIDs
}
\description{
A vector of gene_ids used to create 'sampleGeneSet' example
}
\usage{data(sampleData)}
\format{
  The format is:
 chr "sampleIDs"
}
\details{
None
}
\source{
None
}
\references{
None
}
\examples{
data(sampleData)

}
\keyword{datasets}
